+++ projects/angular2-smart-table/src/lib/components/filter/filter-types/select-filter.component.ts
@@ -3,7 +3,6 @@ import {NgControl} from '@angular/forms';
 import {debounceTime, distinctUntilChanged, skip} from 'rxjs/operators';
 
 import {DefaultFilter} from './default-filter';
-import {defaultStringEqualsFilter, defaultStringInclusionFilter} from "../../../lib/data-source/local/local.filter";
 import {FilterSettings, ListFilterSettings} from "../../../lib/settings";
 
 @Component({
@@ -28,8 +27,10 @@ export class SelectFilterComponent extends DefaultFilter implements OnInit {
   config!: ListFilterSettings;
 
   ngOnInit() {
-    this.column.filterFunction = this.onFilterValues.bind(this);
     this.config = (this.column.filter as FilterSettings).config as ListFilterSettings;
+    if (this.config.strict === undefined || this.config.strict) {
+      this.column.filterFunction = (v, f) => v === f;
+    }
 
     const exist = this.inputControl.valueChanges;
     if (!exist) {
@@ -42,13 +43,5 @@ export class SelectFilterComponent extends DefaultFilter implements OnInit {
         debounceTime(this.delay)
       )
       .subscribe((value: string) => this.setFilter());
-  }
-
-  onFilterValues(cellValue: string, search: string) {
-    if (this.config.strict === undefined || this.config.strict) {
-      return defaultStringEqualsFilter(cellValue, search);
-    } else {
-      return defaultStringInclusionFilter(cellValue, search);
-    }
   }
 }